refactor(agent): extract frontmatter guard to validator, add DEPRECATED classification
- Move checkFrontmatterPreservation and extractFrontmatterField from
  skill-tools.ts to skill-validator.ts so they're reusable by tests
  and any future tool that writes SKILL.md files
- Add DEPRECATED classification to skill-evolve assess phase for
  skills whose underlying procedure is no longer relevant
- Evolve phase retires deprecated skills (update status + delete from disk)
